Add files via upload
This paper introduces a supersymmetric dual-matter atomic model based on two intersecting fields that periodically vary in either the same or opposite phases, forming a shared nucleus of two transversal and two vertical subfields that represent the particles and antiparticles of the dual atomic nucleus. The bosonic or fermionic characteristics of the nuclear subfields are determined by their topological transformations, which are caused by the pushing forces generated by the negative or positive curvatures of the intersecting fields during their contraction or expansion, and by the periodical synchronization and desynchronization of the phases of the intersecting fields while rotating. 
With a mainly visual and conceptual approach, the model employs a set of 2x2 complex rotational matrices of eigenvectors related in a modular way to Sobolev interpolations and to Tomita-Takesaki theory, graphically illustrating reflection positivity, the mass gap problem, or the arising of a purely imaginary time dimension, between other topics.  
The article first explains the fields model in a general way, then it introduces some mathematical formalisms, translates the general system to the standard atomic terminology, and finally compares the model with already known developments and theories.